Break apart the OTP settings styles
